# src/main/java/com/juansenen/gaticket/doc/Gaticket_API_doc.yaml
openapi: 3.0.1
info:
  title: GATICKET API
  version: "1.0.0"
  description: API diseñada para el control y consulta de la web app Gaticket. Gestor de incidencias técnicas, al cual pueden conectarse los usuarios para poder comunicar al equipo de Tecnologias de la Información de la empresa las incidencias surgidas en sus equipos; y por otra parte los administradores pueden recibirlas, gestionarlas y llevar un control sobre la base de datos de los dispositivos a su cargo.

    Para poder realizar un monitoreo y control de la API, se ha incuido la libreria spring-doc actuator. Accesible en el siguiente link

    Todos los endppoints de actuator [API actuator](http://localhost:8080/actuator)
    endpoints de metricas [API actuator](http://localhost:8080/actuator/metrics)
servers:
  - url: http://localhost:8080
    description: Generated server url
...
paths:
...
  /net:
    get:
      tags:
        - net
      summary: Retrieve all network data
      description: Get all data about network.
      operationId: getAll
      responses:
        "400":
          description: Invalid
          content: {}
        "404":
          description: Not Found
          content:
            application/hal+json:
              schema:
                $ref: '#/components/schemas/ErrorMessage'
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Net'
    post:
      tags:
        - net
      summary: Save a new net data
      description: Save anew net data.
      operationId: addOne
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Net'
        required: true
      responses:
        "400":
          description: Invalid
          content: {}
        "404":
          description: Bad Request
          content: {}
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Net'
  /net/{idNet}/{idDevice}:
    post:
      tags:
        - net
      summary: Assing a new net data to a device
      description: Assing net data to a device.
      operationId: addDeviceNet
      parameters:
        - name: idNet
          in: path
          description: Id of net
          required: true
          schema:
            type: integer
            format: int64
        - name: idDevice
          in: path
          description: Id of a device
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "400":
          description: Invalid
          content: {}
        "404":
          description: Bad Request
          content: {}
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Net'
...
  /net/{idNet}:
    delete:
      tags:
        - net
      summary: Free a net by Id
      description: Free a type by Id
      operationId: deteleIp
      parameters:
        - name: idNet
          in: path
          description: Id of the net ip
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "400":
          description: Bad Request
          content:
            application/hal+json:
              schema:
                $ref: '#/components/schemas/ErrorMessage'
        "404":
          description: Not Found
          content:
            application/hal+json:
              schema:
                $ref: '#/components/schemas/ErrorMessage'
        "200":
          description: OK
    patch:
      tags:
        - net
      summary: Update status of net ip
      description: Update status of net ip
      operationId: updateNet
      parameters:
        - name: idNet
          in: path
          description: Id of the net ip
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Net'
        required: true
      responses:
        "400":
          description: Bad Request
          content:
            application/hal+json:
              schema:
                $ref: '#/components/schemas/ErrorMessage'
        "404":
          description: Not Found
          content:
            application/hal+json:
              schema:
                $ref: '#/components/schemas/ErrorMessage'
        "200":
          description: OK
          content:
            application/hal+json:
              schema:
                $ref: '#/components/schemas/Net'
...
  /net/department/{idNet}:
    get:
      tags:
        - net
      summary: Retrieve all network data
      description: Get all data about network.
      operationId: getDepartmentByIp
      parameters:
        - name: idNet
          in: path
          description: Id of net
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "400":
          description: Invalid
          content: {}
        "404":
          description: Not Found
          content:
            application/hal+json:
              schema:
                $ref: '#/components/schemas/ErrorMessage'
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Net'
...
components:
  schemas:
    ErrorMessage:
      type: object
      properties:
        code:
          type: integer
          format: int32
        message:
          type: string
        errors:
          type: object
          additionalProperties:
            type: string
...
    Net:
      type: object
      properties:
        netId:
          type: integer
          description: netId
          format: int64
          readOnly: true
          example: 8
        netIp:
          type: string
          description: IP
          example: 10.53.3.456
        netMask:
          type: string
          description: Mask of network
          example: 255.255.255.0
        netCdir:
          type: string
          description: CDIR number
          example: "24"
        netGateWay:
          type: string
          description: Gateway number
          example: 10.52.2.1
        netStatus:
          type: boolean
          description: Status of the IP
          example: false
      description: Network information of a device
...